fast-api-tutorial/jwt-auth/requirements.txt
fastapi==0.100.0
uvicorn==0.23.2

PyJWT==2.8.0
python-decouple==3.8